pyproject.toml
[project]
name = "zy_cube"
version = "0.1.0"
description = "ZY_Cube 卓越魔方 - 动态可扩展的AI开发平台"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "pillow>=12.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.42.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["app"]
